google gemini LLM/requirements.txt
streamlit
PyPDF2
google.generativeai